infra: Add pull request presubmit checks, continuous `main` checks (#3)
This change adds a Github workflow to run `go test ./...` and `bazel
test //...` on each commit to `main` and `release` branch (should any
exist) as well as PR reviews. A bazel config is created to customize
bazel's behavior:
* keep_going (run all tests possible, even if build fails)
* UI conveniences (show timestamps, no curses, etc.)

The first iteration of these checks are basic actions that don't use our
custom CI runners or RBE clusters, both of which can be added/enabled in
future changes. As such, some amount of extra effort is taken to
preserve bazel's local cache as much as possible, to speed up checks.

Bug: linear/CUS-345