automation: Add back Golang builds (#8)
This change adds back builds using the Go toolchain, which can be
downloaded and run via `bazel run`. Obtaining the Go toolchain this way
saves us from having to install it on the self-hosted runners directly,
and keeps the toolchain used in sync with the one used by bazel
automatically.

The matrix builds are moved to using the Go toolchain, and building
using bazel is done only with Linux. The rationale is:
* building on all platforms is useful for test coverage, and need only
be done using either build system, but not necessarily both
* third parties are more likely to build the project using the Go
tooling
* we intend to ship bazel-built binaries, so we need at least one build
using bazel to protect the release builds

Bug: CUS-345