release: Stage artifacts in temp dir during release
This change resolves a failure to create releases because the artifacts
being uploaded have the same basename after resolving all symlinks,
which the github CLI/Github API is unable to handle.

To resolve this, artifacts are staged with a unique name (matching the
bazel target name) in a temp directory, effectively resolving symlinks;
the unique names are then passed to the Github CLI, which is able to
create the release with all artifacts.

This change also adds beginning + end log messages on logical groups of
steps, to provide additional feedback during the release, and obviating
the need for some comments.

Bug: linear/CUS-332