infra: Add copyright headers check
This change adds:

* a Go tool to check for copyright headers
* a Github job to run the check in presubmit and postsubmit
* copyright headers to all files, to satisfy the check

Because this check necessarily depends on all source files, it can't be
implemented as a bazel test, so it is instead designed to be invoked via
`bazel run` (depends on an environment var that locates the workspace
root). Go was chosen over bash to avoid complicated pipelines and
array/associative array bash syntax.