infra: Add copyright headers check, copyright headers (#18)
This change adds:

* a Go tool to check for copyright headers
* a Github job to run the check in presubmit and postsubmit
* copyright headers to all files, to satisfy the check

Because this check necessarily depends on all source files, it can't be
implemented as a bazel test, so it is instead designed to be invoked via
`bazel run` (depends on an environment var that locates the workspace
root). Go was chosen over bash to avoid complicated pipelines and
array/associative array bash syntax.

Tested: Wrote the tool first; it identified all the source files in the
repo, which were resolved either by adding a header, or by adding an
exception inside the `generatedFiles` var.