buildstamp: Offer version string via function call
Currently, version info is available as part of the version information
returned pre-formatted for printing to a console/log. However, we want
to get version info to share with an external server, so this change
creates such a function.

If the binary has no stamped version info, the function will error
(which implies that during development, testing against a real server
will require version stamping to be enabled).

If the released binary is a real release, it will return a common
version string (`vX.Y.Z`). If the binary is not a real release, it will
return a pseudo-version similar to the format used for [Go
modules](https://go.dev/ref/mod#pseudo-versions) - namely,
`v0.0.0-yyyymmddhhmmss-abcdefabcdef`, where the date is the build date,
and the commit is the commit the tool's build was based on. No component
of the version string will indicate if the code tree was dirty.

Tested: unit tests only
Bug: linear/CUS-385