CUS-384: engflow_auth: support storing tokens in plaintext files
The login and import subcommands now support the -store flag, which
may be set to either 'keyring' (default) or 'file'. 'keyring' is
the same as before. When -store=file, engflow_auth stores tokens
as plaintext JSON files in the user's configuration directory.

Since different tokens may now be stored in different places,
we now maintain a JSON configuration file, tracking which -store value
was used to write each token. This could be used in the future to
implement 'engflow_auth logout -all' or to track token aliases.

File storage is needed in non-interactive environments when no
encrypted keyring is available. On Linux, the go-keyring package
we use relies on dbus-launch, which is installed with GNOME.
We don't expect it to be available on CI machines.