main: Do initialization after flag parsing (#36)
Currently, the selected implementations to any of the internal
interfaces are constructed before CLI arg parsing occurs, which means
that CLI args cannot change which implementations are used. This has
been sufficient until now, but limits our ability to e.g. have a flag
that controls the token storage location.

This change defers initialization until after arg parsing, so that it
can depend on flags added in the future.

This is done by moving the creation of implementations into a method on
`*appState`, which is attached to the CLI app as a `Before` action. This
method only replaces implementations that aren't pre-populated, so that
tests have the opportunity to pre-populate `appState` with fakes.

Tested: unit tests still pass; `engflow_auth login` to dev cluster still
works