CUS-384: fix -store flag, make test less fake-reliant
In #37, we thought App.Before was called after subcommand flag parsing,
but it is not. This meant the -store flag did not work. We need the flags
to be parsed before we construct the object graph in appState.build.
Instead, call appState.build from the Command.Before hook.

Our tests in main_test.go should have caught this, but they fake
appState.tokenStore so the -store flag is ignored. This PR now makes
the test use the real LoadStorer implementations in most cases, but
calls keyring.MockInit ahead of time and uses a temporary user configuration
directory to avoid side effects.