REC-102: release: multiple fixes (#56)
Set uploaded artifact names uniquely so that different jobs upload
separate artifacts. The default was to overwrite the same artifact. No
such setting is needed for downloading: the default is to download all
artifacts.

Fixed the regular expression used to validate semantic versions. We used
a PCRE-compatible regex from semver.org, but grep doesn't use PCRE, and
different versions of grep (on Linux and macOS at least) interpret the
same expression differently. Hopefully we've landed on something that
works.

Added an .exe extension to the Windows artifact.