REC-110: don't use keyring when an unencrypted file token is present
On Linux, the keyring library can prompt the user for a password if
it's not automatically unlocked. This causes engflow_auth to hang
when invoked as a credential helper because neither stdin nor stdout
are connected to a terminal.

With this change, the get command (and anything else that calls loadToken)
now checks for a token created with -store=file first, then falls back
to the keyring library if that's not present. This reverses the previous
order.

When storing a token into the keyring, storeToken now deletes the token
created with -store=file, if present, preventing it from taking precedence.

Together, these changes that mean when -store=file is used, the keyring
library should only be used by the logout command, which attempts to
delete both tokens.