REC-110: refactor to remove CacheAlert (#58)
This is the beginning of a refactoring to move some LoadStorer
implementations into appState.

Although CacheAlert satisfied the LoadStorer interface type,
it was a decorator that didn't actually load or store any tokens
or provide useful or meaningful abstraction. It seems better to
squash this into appState and remove unnecessary abstraction.
This also makes the test a little more realistic.

Also refactored fake token generation to make testing easier.