REC-195: recommend -store=file when keyring is not available (#66)
In the login and import commands, test if we can load a token from
the keyring before doing anything. If the keyring is not available,
print a message that recommends -store=file.

Previously, we could go through the whole login flow then fail to
store the token. This was frustrating for the user since it wasn't
clear what to do next (keyring errors are obscure), and once they
figure it out, then have to go through the login flow again.